refactor(cli): split cli.rs into module structure
Split the monolithic cli.rs (2842 lines) into a cleaner module structure:
- cli/mod.rs (1887 lines): Cli, Commands, StepCommand, shared utilities
- cli/config.rs (748 lines): Config-related enums (ConfigCommand, StateCommand, etc.)
- cli/hook.rs (208 lines): HookCommand
- cli/list.rs (18 lines): ListSubcommand

No functional changes - all 738 integration tests and 406 unit tests pass.
